add correct mode to index-events, support polymer ibc dispatcher v2

+++ src/graphql/typeDefs.js
@@ -130,7 +130,7 @@ module.exports = /* GraphQL */ `
     sequence: Int
     from: Address
     currentState: PacketType
-    timeout: Int
+    timeout: Float
     timestamp: Int
     states: [PacketState]
   }
